Change machine id to 16bits, change CAN packet format

+++ Inc/can_io.h
@@ -5,14 +5,14 @@
 
 extern uint32_t can_tx_mailbox_used;
 
-typedef void(*CAN_CmdListener)(uint8_t* data, uint8_t len, uint8_t from, _Bool isBroadcast);
+typedef void(*CAN_CmdListener)(uint8_t* data, uint8_t len, uint16_t from, _Bool isBroadcast);
 typedef void(*CAN_ReplyListener)(_Bool ok, uint8_t from);
 
-void can_init(uint8_t selfid);
+void can_init(uint16_t selfid);
 void can_set_cmdlistener(CAN_CmdListener lis);
 void can_set_replylistener(CAN_ReplyListener lis);
-void can_set_id(uint8_t id);
-HAL_StatusTypeDef can_send_cmd(uint8_t *data, uint8_t len, uint8_t to);
+void can_set_id(uint16_t id);
+HAL_StatusTypeDef can_send_cmd(uint8_t *data, uint8_t len, uint16_t to);
 HAL_StatusTypeDef can_send_reply(uint8_t to, _Bool ok);
